all the changes are tied with relocation components from general directory. Code base hasn't been chaged
