+++ src/app/app.module.ts
@@ -4,12 +4,14 @@ import { AppRoutingModule } from './app-routing.module';
 import { HttpClientModule } from '@angular/common/http';
 import { SharedModule } from './shared/shared.module';
 import { CoreModule } from './core.module';
-import { AuthModule } from './components/auth/auth.module';
+import { AuthModule } from './auth/auth.module';
 
 import { AppComponent } from './app.component';
-import { HeaderComponent } from './components/header/header.component';
+import { HeaderComponent } from './header/header.component';
 import { StoreModule } from '@ngrx/store';
-import { shoppingListReducer } from './components/shopping-list/store/shopping-list.reducer';
+import { shoppingListReducer } from './shopping-list/store/shopping-list.reducer';
+import { authReducer } from './auth/store/auth.reducer';
+
 
 
 
@@ -27,7 +29,12 @@ import { shoppingListReducer } from './components/shopping-list/store/shopping-l
       SharedModule,
       CoreModule,
       AuthModule,
-      StoreModule.forRoot({shoppingList: shoppingListReducer}) //here we tell to ngrx which reducers we have in our application
+      StoreModule.forRoot(//here we tell to ngrx which reducers we have in our application
+        {
+          shoppingList: shoppingListReducer,
+          auth: authReducer
+        }
+      ) 
     ],
 
   bootstrap: [AppComponent]
