+++ src/app/auth/auth.module.ts
@@ -2,7 +2,7 @@ import { NgModule } from "@angular/core";
 import { FormsModule } from "@angular/forms";
 import { AuthComponent } from "./auth.component";
 import { RouterModule } from "@angular/router";
-import { SharedModule } from "../../shared/shared.module";
+import { SharedModule } from "../shared/shared.module";
 
 
 @NgModule({
