+++ src/app/auth/store/auth.reducer.ts
@@ -1,0 +1,13 @@
+import { User } from "../user.model";
+
+export interface State{
+    user: User;
+}
+
+const initialState: State = {
+    user: null
+}
+
+export function authReducer(state = initialState, action){
+    return state;
+}
