+++ src/app/core.module.ts
@@ -1,8 +1,8 @@
 import { HTTP_INTERCEPTORS } from "@angular/common/http";
 import { NgModule } from "@angular/core";
-import { AuthInterceptorService } from "./components/auth/auth-interceptor.service";
-import { RecipeService } from "./components/recipes/recipe.service";
-import { ShoppingListService } from "./components/shopping-list/shopping-list.service";
+import { AuthInterceptorService } from "./auth/auth-interceptor.service";
+import { RecipeService } from ".//recipes/recipe.service";
+import { ShoppingListService } from "./shopping-list/shopping-list.service";
