+++ src/app/recipes/recipes-resolve.service.ts
@@ -1,7 +1,7 @@
 import { ActivatedRouteSnapshot, Resolve, RouterStateSnapshot } from "@angular/router";
 import { Recipe } from "./recipe.model";
 import { Observable } from "rxjs";
-import { DataStorageService } from "../../shared/data-storage.service";
+import { DataStorageService } from "../shared/data-storage.service";
 import { Injectable } from "@angular/core";
 import { RecipeService } from "./recipe.service";
