+++ src/app/recipes/recipes.module.ts
@@ -8,7 +8,7 @@ import { RecipesComponent } from "./recipes.component";
 import { ReactiveFormsModule } from "@angular/forms";
 import { RouterModule } from "@angular/router";
 import { RecipesRoutingModule } from "./recipe-routing.module";
-import { SharedModule } from "../../shared/shared.module";
+import { SharedModule } from "../shared/shared.module";
 
 
 @NgModule({
