+++ src/app/shopping-list/shopping-list-edit/shopping-list-edit.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnDestroy, OnInit, ViewChild } from '@angular/core';
-import { Ingredient } from '../../../shared/ingredient.model';
+import { Ingredient } from '../../shared/ingredient.model';
 //import { ShoppingListService } from '../shopping-list.service';
 import { FormControl, NgForm } from '@angular/forms';
 import { Subscription } from 'rxjs';
