+++ src/app/shopping-list/shopping-list.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from "@angular/core";
 import { ShoppingListRoutingModule } from "./shopping-list-routing.module";
 import { FormsModule } from "@angular/forms";
-import { SharedModule } from "../../shared/shared.module";
+import { SharedModule } from "../shared/shared.module";
 
 import { ShoppingListEditComponent } from "./shopping-list-edit/shopping-list-edit.component";
 import { ShoppingListComponent } from "./shopping-list.component";
