+++ src/app/shopping-list/shopping-list.service.ts
@@ -1,5 +1,5 @@
 import { Subject } from 'rxjs';
-import { Ingredient } from '../../shared/ingredient.model';
+import { Ingredient } from '../shared/ingredient.model';
