+++ src/app/shopping-list/store/shopping-list.actions.ts
@@ -1,5 +1,5 @@
 import { Action } from "@ngrx/store";
-import { Ingredient } from "../../../shared/ingredient.model";
+import { Ingredient } from "../../shared/ingredient.model";
 
 export const ADD_INGREDIENT = 'ADD_INGREDIENT';
 export const ADD_INGREDIENTS = 'ADD_INGREDIENTS';
