+++ src/app/shopping-list/store/shopping-list.reducer.ts
@@ -1,5 +1,5 @@
 import { Action } from "@ngrx/store";
-import { Ingredient } from "../../../shared/ingredient.model";
+import { Ingredient } from "../../shared/ingredient.model";
 import * as ShoppingListActions from "./shopping-list.actions";
 
 export interface State {
